Separate implicit return mixin
